pyproject.toml
[project]
name = "Requirement-analyer"
version = "0.1.0"
description = ""
authors = [
    {name = "Ravindra Shinde"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "^1.45.1"
langchain-openai = "^0.3.18"
langchain-community = "^0.3.24"
pymupdf = "^1.26.0"
python-docx = "^1.1.2"
python-dotenv = "^1.1.0"
faiss-cpu = "^1.11.0"
sentence-transformers = "^4.1.0"
langchain-huggingface = "^0.2.0"


[tool.poetry.group.dev.dependencies]
certifi = "^2025.4.26"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
